# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prediction Game API",
    "description": "API for wallet-authenticated prediction gameplay, leaderboards, rounds, and predictions. Use Swagger UI to explore endpoints and test requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste a JWT like: Bearer <token>"
      }
    },
...
  },
...
  "paths": {
...
    "/api/admin/dead-letter": {
      "get": {
        "summary": "List failed notification/event dispatches",
        "description": "Returns the dead-letter queue contents, newest first. Admin only.\nUse `?status=` and `?channel=` to filter; defaults to all rows.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/dead-letter/retry-all": {
      "post": {
        "summary": "Replay every pending/retrying dispatch in the DLQ",
        "description": "Admin only. Returns a counts summary.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/dead-letter/{id}/retry": {
      "post": {
        "summary": "Replay a single DLQ entry",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}